+++ NEWS.md
@@ -4,3 +4,4 @@
 # SDMtune 0.1.0.9000
 * Fix bug in VarImp function for SDMmodelCV objects
 * Real-time charts are now displayed also when R is not executed from RStudio
+* `maxentVarImp` function available for `SDMmodelCV` objects